@@ -23,7 +23,7 @@
     "@babel/core": "^7.5.5",
     "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
     "@babel/preset-env": "^7.5.5",
-    "@vue/test-utils": "^1.0.0-beta.29",
+    "@vue/compiler-sfc": "npm:@vue/compiler-sfc@^3",
     "autoprefixer": "^9.6.1",
     "babel-loader": "^9.1.2",
     "browserslist": "^4.6.6",
@@ -39,10 +39,8 @@
     "mini-css-extract-plugin": "^2.7.6",
     "postcss-loader": "^3.0.0",
     "sass": "^1.63.6",
-    "sass-loader": "^13.3.2",
+    "sass-loader": "^14.2.1",
     "style-loader": "^0.23.1",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.10",
     "webpack": "^5.88.1",
     "webpack-bundle-analyzer": "^4.9.0",
     "webpack-bundle-tracker": "^2.0.1",
@@ -59,9 +57,9 @@
     "popper.js": "^1.15.0",
     "query-string": "^6.12.1",
     "slick-carousel": "^1.8.1",
-    "vue": "^2.7.14",
-    "vue-loader": "^15.10.1",
-    "vue-select": "^3.4.0",
+    "vue": "^3.4.24",
+    "vue-loader": "^17.4.2",
+    "vue-select": "^4.0.0-beta.6",
     "vue2-dropzone": "^3.6.0"
   }
 }
