@@ -1,10 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import CompetenceFilterApp from './CompetenceFilterApp.vue'
 
 if (document.getElementById('filter-app')) {
-  new Vue({
-    components: {CompetenceFilterApp},
-    template: '<CompetenceFilterApp/>',
-    render: h => h(CompetenceFilterApp)
-  }).$mount('#filter-app')
+  createApp(CompetenceFilterApp).mount('#filter-app')
 }
