@@ -1,10 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import OverviewApp from './OverviewApp.vue'
 
 if (document.getElementById('overview-app')) {
-  new Vue({
-    components: {OverviewApp},
-    template: '<OverviewApp/>',
-    render: h => h(OverviewApp)
-  }).$mount('#overview-app')
+  createApp(OverviewApp).mount('#overview-app')
 }
