@@ -1,11 +1,12 @@
 import $ from 'jquery'
-import Vue from 'vue'
 import Tooltip from 'bootstrap/js/dist/tooltip'
 
-export const tooltipDirective = Vue.directive('tooltip', function(el, binding){
+export const tooltipDirective = {
+  mounted() {
     if (!el) return;
     new Tooltip($(el), {
       title: binding.value,
       trigger: 'hover'
     })
-  })
+  }
+};
